requirements.txt
fastapi
uvicorn
httpx
firebase-admin
python-dotenv
python-jose[cryptography]
passlib